refactor: seperate codec into different files
