+++ src/core/metrics.c
@@ -230,10 +230,9 @@ static void on_v2ray_ws_g204_read(struct bufferevent *bev, void *ctx)
 				connect_time;
 			pgs_session_t dummy = { 0 };
 			dummy.outbound = mctx->outbound;
-			uint64_t total_len = pgs_vmess_write_remote(
+			size_t total_len = pgs_vmess_write_remote(
 				&dummy, (const uint8_t *)g204_http_req,
-				strlen(g204_http_req),
-				(pgs_session_write_fn)&vmess_flush_remote);
+				strlen(g204_http_req));
 		}
 	} else {
 		double g204_time = elapse(mctx->start_at);
@@ -313,10 +312,10 @@ static void on_v2ray_tcp_g204_event(struct bufferevent *bev, short events,
 		pgs_session_t dummy = { 0 };
 		dummy.outbound = mctx->outbound;
 		struct evbuffer *output = bufferevent_get_output(bev);
-		uint64_t total_len = pgs_vmess_write_remote(
-			&dummy, (const uint8_t *)g204_http_req,
-			strlen(g204_http_req),
-			(pgs_session_write_fn)&vmess_flush_remote);
+		size_t total_len =
+			pgs_vmess_write_remote(&dummy,
+					       (const uint8_t *)g204_http_req,
+					       strlen(g204_http_req));
 		pgs_logger_debug(mctx->logger, "g204 req sent: %d", total_len);
 	}
 	if (events & BEV_EVENT_ERROR)
